requirements.txt
Django==1.4.1
MySQL-python==1.2.3
PIL==1.1.7
South==0.7.6
-e hg+http://bitbucket.org/ubernostrum/django-contact-form//@1d3791fa4dfbb2375ba0d27e62283380c0ec74e6#egg=django_contact_form-dev
django-tinymce==1.5.1b2
wsgiref==0.1.2
